extract `getDataForSearchInput`
